feat: Migrate money fields to long type in database schema (#79)
- Added migration to change money fields (GrandTotal, TaxAmount, Subtotal, ShippingFee, DiscountAmount) from decimal to long in fulfillment_saga_states and checkout_saga_states tables.
- Updated OrderDbContextModelSnapshot to reflect changes in the database schema.
- Modified CheckoutSagaState and FulfillmentSagaState to use long for monetary values.
- Introduced new FulfillmentSagaState and its entity configuration for handling fulfillment processes.
- Enhanced GetCheckoutStatusQueryHandler to throw NotFoundException if the saga is not found.
- Updated OrderInfrastructureExtension to register event publisher services.
- Adjusted Login page to handle registration intent from authorization context.
- Changed DefaultRedirectUrl in appsettings.json for UserService.

+++ libs/HiveSpace.Infrastructure.Messaging.Shared/CheckoutSaga/Commands/CreateOrder.cs
@@ -8,10 +8,10 @@ public record CreateOrder
     public Guid               UserId          { get; init; }
     public List<OrderItemDto> Items           { get; init; } = new();
     public DeliveryAddressDto DeliveryAddress { get; init; } = null!;
-    public decimal            Subtotal        { get; init; }
-    public decimal            ShippingFee     { get; init; }
-    public decimal            TaxAmount       { get; init; }
-    public decimal            DiscountAmount  { get; init; }
-    public decimal            GrandTotal      { get; init; }
+    public long               Subtotal        { get; init; }
+    public long               ShippingFee     { get; init; }
+    public long               TaxAmount       { get; init; }
+    public long               DiscountAmount  { get; init; }
+    public long               GrandTotal      { get; init; }
     public PaymentMethod      PaymentMethod   { get; init; } = PaymentMethod.COD;
 }
